libpkg: repo: encode the signature type
The signature type is added to each archive that's signed.  For backwards
compatibility, it's added to its own signature_type file rather than
prepending to the signature as we do elsewhere.

This lets the client complain gracefully if they're presenting a public
key of the wrong type, and removes the need to guess or ask the client for
the type of key.